requirements.txt
GitPython==3.1.30
PyDriller==2.5
networkx==3.2.1
python-louvain==0.16
numpy==1.26.2
pandas==2.1.3
fastapi==0.110.0
uvicorn==0.24.0
SQLAlchemy==2.0.23
python-dotenv==1.0.0
psycopg2-binary==2.9.9
